Some cleanup, Added description comments
