+++ ObjectPrinting/Serializers/DelegatePropertySerializer.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using ObjectPrinting.Configuration;
 
 namespace ObjectPrinting.Serializers
 {
