+++ ObjectPrinting/Serializers/IPropertySerializer.cs
@@ -1,5 +1,8 @@
-﻿namespace ObjectPrinting.Configuration
+﻿namespace ObjectPrinting.Serializers
 {
+    /// <summary>
+    /// General property serializer
+    /// </summary>
     public interface IPropertySerializer<in TProperty>
     {
         string Serialize(TProperty value);
