+++ ObjectPrinting/Serializers/IgnorePropertySerializer.cs
@@ -1,6 +1,4 @@
-﻿using ObjectPrinting.Configuration;
-
-namespace ObjectPrinting.Serializers
+﻿namespace ObjectPrinting.Serializers
 {
     public class IgnorePropertySerializer<TProperty> : IPropertySerializer<TProperty>
     {
